+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Included workflow fields in OpenAPI document for filtering (OSIDB-2083)
 - Set migrated/duplicated delegated resolution to not affected (OSIDB-1406)
 - Update valid affectedness-resolution combinations (OSIDB-2143)
+- Change Flaw API filter to allow a list of workflow_state (OSIDB-2208)
 
 ### Fixed
 - Fix Jira sync when bugzilla token is present (OSIDB-2171)